+++ CHANGELOG.md
@@ -19,6 +19,7 @@ make sure to check and update your HELM settings.
 * Fixed a typo in the helm variable rewriteTarget.
 * Fixed security queries running as normal user, resulting in too narrow access.
 * Improved the memory efficiency of the DataArray resultFormat.
+* Added checks for maximum username and password lengths.
 
 
 ## Release version 2.2.0